refactored to allow deletion to invalid state also for text fields
